appwrite/functions/process-source/requirements.txt
appwrite
newspaper3k
pymongo
beautifulsoup4
nltk
lxml_html_clean
croniter
google-generativeai